+++ Kokoro/Features/CardRenderManager.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.Xna.Framework;
+﻿using daisyowl.text;
+using Microsoft.Xna.Framework;
 using Shockah.Shared;
 using System.Collections.Generic;
 using System.Linq;
@@ -9,6 +10,14 @@ public sealed class CardRenderManager : HookManager<ICardRenderHook>
 {
 	public bool ShouldDisableCardRenderingTransformations(G g, Card card)
 		=> Hooks.Any(h => h.ShouldDisableCardRenderingTransformations(g, card));
+
+	public Font? ReplaceTextCardFont(G g, Card card)
+	{
+		foreach (var hook in Hooks)
+			if (hook.ReplaceTextCardFont(g, card) is { } font)
+				return font;
+		return null;
+	}
 
 	public Vec ModifyTextCardScale(G g, Card card)
 		=> Hooks.Aggregate(Vec.One, (v, hook) => v * hook.ModifyTextCardScale(g, card));